FlaskBackend/requirements.txt
Flask==2.2.2
numpy==1.23.4
Pillow==9.3.0
tensorflow==2.10.0
